\name{NetPreProc-package}
\alias{NetPreProc-package}
%\alias{NetPreProc}
\docType{package}
\title{ NetPreProc }
\description{
NetPreProc: Network Pre-Processing and normalization.
}
\details{
\tabular{ll}{
Package: \tab NetPreProc\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2011-12-26\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
Depends: \tab methods\cr
Imports: \tab  graph\cr
Suggests: \tab  bionedata\cr
}
Package to pre-process and normalize data in graphs.
}
\author{
Giorgio Valentini, Matteo Re -- Universita' degli Studi di Milano

Maintainer: Giorgio Valentini<valentini@dsi.unimi.it>
}

\keyword{ package }
