\name{NetPreProc-package}
\alias{NetPreProc-package}
%\alias{NetPreProc}
\docType{package}
\title{ NetPreProc }
\description{
Network Pre-Processing and normalization. 
}
\details{
\tabular{ll}{
Package: \tab NetPreProc\cr
Type: \tab Package\cr
Version: \tab 1.2\cr
Date: \tab 2022-09-14\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
Depends: \tab methods\cr
Imports: \tab  graph\cr
Suggests: \tab  bionedata\cr
}
Network Pre-Processing and normalization. Methods for normalizing graphs, including Chua normalization, Laplacian normalization, Binary magnification, min-max normalization and oters. Methods to sparsify adjacency matrices. Methods for graph pre-processing and for filtering edges of the graph.
}
\author{
Giorgio Valentini -- Universita' degli Studi di Milano

Maintainer: Giorgio Valentini<valentini@di.unimi.it>
}

\keyword{ package }
